refac: Move Event related components into a separate folder
